build: TIO-318: Update Rust, toolchain, Edition, FRB side, Cargo minor and bugfix versions (#452)
**Are there specific parts that the reviewer should look out for?**

- [ ] no
- [x] yes (please specify)

This PR updates Rust to the latest version with edition support. Details
are listed in the new documentations.
The following things are updated:
- Rust (1.84.0 to 1.85.0)
- Rust toolchain (1.84.0 to 1.85.0)
- Edition 2021 to Edition 2024
- FRB side code
- multiple Rust minor and bugfix versions (Cargo.toml file)

Updating Rust Edition to 2024 leads to linting and style issues. One
important change is that it wasn't allowed anymore to use the filename
api.rs in the Rust api folder (rust/src/api/api.rs). The file was
renamed to ffi.rs (ffi is the recommended name), which led to
re-generating multiple files.

**Any other comments?** _(e.g., unrelated minor changes piggybacking
this PR)_

- [x] no
- [ ] yes (please specify)

**Helpful screenshots?**

- [x] no
- [ ] yes

---------

Co-authored-by: Jerome Weiß <jerome@cultivate.team>